requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework_simplejwt==4.4.0
requests
PyJWT==1.7.1
